+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 0.2.0
 
-* Rename `setConfig` to `demand`.
-* Rename `getConfig` to `learn`.
+* Rename `set/setConfig` methods to `demand`.
+* Rename `get/getConfig` methods to `learn`.
+* Rename `grunt-horde:set-config` event to `grunt-horde:demand`.
 * fix(require): Resolve paths before existence checks
 * fix(initConfig): setConfig() k/v pairs were not merged